refactor: untangle settings and repository boundaries
