+++ lib/page/setting/setting.dart
@@ -4,1004 +4,41 @@
 
 // Setting window.
 
-import 'dart:async';
-import 'dart:io';
-
-import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
-import 'package:get_it/get_it.dart';
-import 'package:signals/signals_flutter.dart';
 import 'package:styled_widget/styled_widget.dart';
-import 'package:talker_flutter/talker_flutter.dart';
-import 'package:watermeter/controller/update_notice_controller.dart';
-import 'package:watermeter/external/ruisi_flutter/controller/ruisi_controller.dart';
-import 'package:watermeter/model/xidian_ids/classtable.dart';
-import 'package:watermeter/page/homepage/info_widget/classtable_card.dart';
-import 'package:watermeter/page/public_widget/context_extension.dart';
-import 'package:watermeter/page/public_widget/re_x_card.dart';
-import 'package:watermeter/page/setting/dialogs/change_color_dialog.dart';
-import 'package:watermeter/page/setting/dialogs/change_localization_dialog.dart';
-import 'package:watermeter/page/setting/dialogs/schoolnet_password_dialog.dart';
-import 'package:watermeter/page/setting/dialogs/semester_switch_dialog.dart';
-import 'package:watermeter/page/setting/dialogs/update_dialog.dart';
-import 'package:watermeter/page/setting/notification_page/notification_debug_page.dart';
-import 'package:watermeter/page/setting/notification_page/notification_page.dart';
-import 'package:watermeter/repository/localization.dart';
-import 'package:watermeter/repository/logger.dart';
-import 'package:watermeter/page/public_widget/toast.dart';
-import 'package:restart_app/restart_app.dart';
-import 'package:sn_progress_dialog/progress_dialog.dart';
-import 'package:watermeter/controller/classtable_controller.dart';
-import 'package:watermeter/controller/exam_controller.dart';
-import 'package:watermeter/controller/other_experiment_controller.dart';
-import 'package:watermeter/controller/physics_experiment_controller.dart';
-import 'package:watermeter/controller/theme_controller.dart';
-import 'package:watermeter/page/setting/dialogs/experiment_password_dialog.dart';
-import 'package:watermeter/repository/pick_file.dart';
-import 'package:watermeter/repository/preference.dart' as preference;
-import 'package:watermeter/repository/system_calendar_sync_service.dart';
-import 'package:watermeter/page/setting/dialogs/sport_password_dialog.dart';
-import 'package:watermeter/page/setting/dialogs/change_swift_dialog.dart';
-import 'package:watermeter/repository/network_session.dart';
-import 'package:watermeter/repository/user_defined_class_file.dart';
-import 'package:watermeter/repository/xidian_ids/classtable_session.dart';
-import 'package:watermeter/repository/xidian_ids/energy_session.dart';
-import 'package:watermeter/repository/xidian_ids/exam_session.dart';
-import 'package:watermeter/repository/xidian_ids/score_session.dart';
-import 'package:watermeter/repository/xidian_ids/sysj_session.dart';
-import 'package:watermeter/repository/physics_experiment_session.dart';
-import 'package:watermeter/repository/xidian_sport_session.dart';
-import 'package:watermeter/repository/widget_state_sync.dart';
-import 'package:watermeter/themes/color_seed.dart';
-import 'package:watermeter/routing/routes.dart';
+import 'package:watermeter/controller/setting_actions_controller.dart';
+import 'package:watermeter/page/setting/setting_sections.dart';
 
 class SettingWindow extends StatefulWidget {
   const SettingWindow({super.key});
+
   @override
   State<SettingWindow> createState() => _SettingWindowState();
 }
 
 class _SettingWindowState extends State<SettingWindow> {
-  Widget _buildListSubtitle(String text) => Text(
-    text,
-    style: const TextStyle(fontWeight: FontWeight.bold),
-  ).padding(bottom: 8).center();
+  final SettingActionsController _actions = const SettingActionsController();
 
-  bool get _isSemesterAwareControllerLoading =>
-      ClassTableController.i.schoolClassTableStateSignal.value.isLoading ||
-      ExamController.i.examInfoStateSignal.value.isLoading ||
-      PhysicsExperimentController
-          .i
-          .physicsExperimentStateSignal
-          .value
-          .isLoading ||
-      OtherExperimentController.i.otherExperimentStateSignal.value.isLoading;
-
-  Future<void> _waitForSemesterAwareReloads() async {
-    await Future<void>.delayed(const Duration(milliseconds: 100));
-    final stopwatch = Stopwatch()..start();
-    while (_isSemesterAwareControllerLoading &&
-        stopwatch.elapsed < const Duration(seconds: 30)) {
-      await Future<void>.delayed(const Duration(milliseconds: 100));
+  void _refresh() {
+    if (mounted) {
+      setState(() {});
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    List<String> demoBlueModeName = [
-      FlutterI18n.translate(
-        context,
-        "setting.change_brightness_dialog.follow_setting",
-      ),
-      FlutterI18n.translate(
-        context,
-        "setting.change_brightness_dialog.day_mode",
-      ),
-      FlutterI18n.translate(
-        context,
-        "setting.change_brightness_dialog.night_mode",
-      ),
-    ];
     return Scaffold(
       body: ListView(
         padding: const EdgeInsets.all(16),
         children: [
-          Text.rich(
-            TextSpan(
-              children: [
-                TextSpan(
-                  text: Platform.isIOS || Platform.isMacOS || Platform.isAndroid
-                      ? "XDYou"
-                      : 'Traintime PDA',
-                  style: const TextStyle(
-                    fontSize: 20,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-                const TextSpan(
-                  text: '\nWritten by BenderBlog Rodriguez and contributors',
-                ),
-              ],
-            ),
-          ).padding(horizontal: 8.0),
+          const SettingHeader(),
           const SizedBox(height: 20),
-          ReXCard(
-            title: _buildListSubtitle(
-              FlutterI18n.translate(context, "setting.about"),
-            ),
-            remaining: const [],
-            bottomRow: Column(
-              children: [
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(
-                      context,
-                      "setting.about_this_program",
-                    ),
-                  ),
-                  subtitle: Text(
-                    FlutterI18n.translate(
-                      context,
-                      "setting.version",
-                      translationParams: {
-                        "version":
-                            "${preference.packageInfo.version}+"
-                            "${preference.packageInfo.buildNumber}",
-                      },
-                    ),
-                  ),
-                  onTap: () => context.pushReplacementNamed(Routes.about),
-                  trailing: const Icon(Icons.navigate_next),
-                ),
-                const Divider(),
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(context, "setting.check_update"),
-                  ),
-                  subtitle: Watch((context) {
-                    final updateState =
-                        UpdateNoticeController.i.updateMessageStateSignal.value;
-                    return Text(
-                      FlutterI18n.translate(
-                        context,
-                        "setting.latest_version",
-                        translationParams: {
-                          "latest":
-                              updateState.value?.code ??
-                              FlutterI18n.translate(context, "setting.waiting"),
-                        },
-                      ),
-                    );
-                  }),
-                  onTap: () {
-                    showToast(
-                      context: context,
-                      msg: FlutterI18n.translate(
-                        context,
-                        "setting.fetching_update",
-                      ),
-                    );
-                    UpdateNoticeController.i.reloadUpdateNoticeInfo().then((
-                      value,
-                    ) async {
-                      if (context.mounted) {
-                        if (UpdateNoticeController
-                            .i
-                            .updateMessageStateSignal
-                            .value
-                            .hasError) {
-                          showToast(
-                            context: context,
-                            msg: FlutterI18n.translate(
-                              context,
-                              "setting.fetch_failed",
-                            ),
-                          );
-                          return;
-                        }
-                        switch (UpdateNoticeController
-                            .i
-                            .isNewVersionAvaliableComputed
-                            .value) {
-                          case null:
-                            showToast(
-                              context: context,
-                              msg: FlutterI18n.translate(
-                                context,
-
-                                "setting.current_testing",
-                              ),
-                            );
-                          case true:
-                            await showDialog(
-                              context: context,
-                              builder: (context) => Watch(
-                                (context) => UpdateDialog(
-                                  updateMessage: UpdateNoticeController
-                                      .i
-                                      .updateMessageStateSignal
-                                      .value
-                                      .value!,
-                                ),
-                              ),
-                            );
-                          case false:
-                            showToast(
-                              context: context,
-                              msg: FlutterI18n.translate(
-                                context,
-                                "setting.current_stable",
-                              ),
-                            );
-                        }
-                      }
-                    });
-                  },
-                  trailing: const Icon(Icons.navigate_next),
-                ),
-              ],
-            ),
-          ),
-          ReXCard(
-            title: _buildListSubtitle(
-              FlutterI18n.translate(context, "setting.ui_setting"),
-            ),
-            remaining: const [],
-            bottomRow: Column(
-              children: [
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(context, "setting.color_setting"),
-                  ),
-                  subtitle: Text(
-                    FlutterI18n.translate(
-                      context,
-                      "setting.change_color_dialog."
-                      "${ColorSeed.values[preference.getInt(preference.Preference.color)].label}",
-                    ),
-                  ),
-                  trailing: const Icon(Icons.navigate_next),
-                  onTap: () {
-                    showDialog(
-                      context: context,
-                      builder: (context) => const ChangeColorDialog(),
-                    );
-                  },
-                ),
-                const Divider(),
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(
-                      context,
-                      "setting.brightness_setting",
-                    ),
-                  ),
-                  subtitle: Text(
-                    demoBlueModeName[preference.getInt(
-                      preference.Preference.brightness,
-                    )],
-                  ),
-                  trailing: ToggleButtons(
-                    isSelected: List<bool>.generate(
-                      3,
-                      (index) =>
-                          index ==
-                          preference.getInt(preference.Preference.brightness),
-                    ),
-                    onPressed: (int value) {
-                      preference
-                          .setInt(preference.Preference.brightness, value)
-                          .then((value) {
-                            setState(() {
-                              ThemeController.i.updateTheme();
-                            });
-                          });
-                    },
-                    children: const [
-                      Icon(Icons.phone_android_rounded),
-                      Icon(Icons.light_mode_rounded),
-                      Icon(Icons.dark_mode_rounded),
-                    ],
-                  ),
-                ),
-                const Divider(),
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(context, "setting.simplify_timeline"),
-                  ),
-                  subtitle: Text(
-                    FlutterI18n.translate(
-                      context,
-                      "setting.simplify_timeline_description",
-                    ),
-                  ),
-                  trailing: Switch(
-                    value: preference.getBool(
-                      preference.Preference.simplifiedClassTimeline,
-                    ),
-                    onChanged: (bool value) {
-                      setState(() {
-                        preference
-                            .setBool(
-                              preference.Preference.simplifiedClassTimeline,
-                              value,
-                            )
-                            .then(
-                              (value) =>
-                                  ClassTableCard.reloadSettingsFromPref(),
-                            );
-                      });
-                    },
-                  ),
-                ),
-                const Divider(),
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(
-                      context,
-                      "setting.localization_dialog.title",
-                    ),
-                  ),
-                  subtitle: Text(
-                    FlutterI18n.translate(
-                      context,
-                      Localization.values
-                          .firstWhere(
-                            (value) =>
-                                value.string ==
-                                preference.getString(
-                                  preference.Preference.localization,
-                                ),
-                          )
-                          .toShow,
-                    ),
-                  ),
-                  trailing: const Icon(Icons.navigate_next),
-                  onTap: () {
-                    showDialog(
-                      barrierDismissible: false,
-                      context: context,
-                      builder: (context) => const ChangeLanguageDialog(),
-                    );
-                  },
-                ),
-              ],
-            ),
-          ),
-          ReXCard(
-            title: _buildListSubtitle(
-              FlutterI18n.translate(context, "setting.account_setting"),
-            ),
-            remaining: const [],
-            bottomRow: Column(
-              children: [
-                if (!preference.getBool(preference.Preference.role)) ...[
-                  ListTile(
-                    title: Text(
-                      FlutterI18n.translate(
-                        context,
-                        "setting.sport_password_setting",
-                      ),
-                    ),
-                    trailing: const Icon(Icons.navigate_next),
-                    onTap: () {
-                      showDialog(
-                        barrierDismissible: false,
-                        context: context,
-                        builder: (context) => const SportPasswordDialog(),
-                      );
-                    },
-                  ),
-                  const Divider(),
-                  ListTile(
-                    title: Text(
-                      FlutterI18n.translate(
-                        context,
-                        "setting.experiment_password_setting",
-                      ),
-                    ),
-                    trailing: const Icon(Icons.navigate_next),
-                    onTap: () {
-                      showDialog(
-                        barrierDismissible: false,
-                        context: context,
-                        builder: (context) => const ExperimentPasswordDialog(),
-                      );
-                    },
-                  ),
-                  const Divider(),
-                ],
-
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(
-                      context,
-                      "setting.schoolnet_password_setting",
-                    ),
-                  ),
-                  subtitle: Text(
-                    FlutterI18n.translate(
-                      context,
-                      "setting.schoolnet_password_description",
-                    ),
-                  ),
-                  trailing: const Icon(Icons.navigate_next),
-                  onTap: () {
-                    showDialog(
-                      context: context,
-                      builder: (context) => const SchoolNetPasswordDialog(),
-                    );
-                  },
-                ),
-              ],
-            ),
-          ),
-          if (Platform.isAndroid || Platform.isIOS)
-            ReXCard(
-              title: _buildListSubtitle(
-                FlutterI18n.translate(context, "setting.notification_setting"),
-              ),
-              remaining: const [],
-              bottomRow: Column(
-                children: [
-                  ListTile(
-                    title: Text(
-                      FlutterI18n.translate(
-                        context,
-                        "setting.course_reminder_setting",
-                      ),
-                    ),
-                    subtitle: Text(
-                      FlutterI18n.translate(
-                        context,
-                        "setting.course_reminder_description",
-                      ),
-                    ),
-                    trailing: const Icon(Icons.navigate_next),
-                    onTap: () {
-                      context.push(const NotificationSettingsPage());
-                    },
-                  ),
-                ],
-              ),
-            ),
-          ReXCard(
-            title: _buildListSubtitle(
-              FlutterI18n.translate(context, "setting.classtable_setting"),
-            ),
-            remaining: const [],
-            bottomRow: Column(
-              children: [
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(context, "setting.background"),
-                  ),
-                  trailing: Switch(
-                    value: preference.getBool(preference.Preference.decorated),
-                    onChanged: (bool value) {
-                      if (value == true &&
-                          !preference.getBool(
-                            preference.Preference.decoration,
-                          )) {
-                        showToast(
-                          context: context,
-                          msg: FlutterI18n.translate(
-                            context,
-                            "setting.no_background",
-                          ),
-                        );
-                      } else {
-                        setState(() {
-                          preference.setBool(
-                            preference.Preference.decorated,
-                            value,
-                          );
-                        });
-                      }
-                    },
-                  ),
-                ),
-                const Divider(),
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(context, "setting.choose_background"),
-                  ),
-                  trailing: const Icon(Icons.navigate_next),
-                  onTap: () async {
-                    PlatformFile? result;
-                    try {
-                      result = await pickFile(type: FileType.image);
-                    } on MissingStoragePermissionException {
-                      if (context.mounted) {
-                        showToast(
-                          context: context,
-                          msg: FlutterI18n.translate(
-                            context,
-                            "setting.no_permission",
-                          ),
-                        );
-                      }
-                    }
-                    if (mounted) {
-                      if (result != null) {
-                        File(result.path!).copySync(
-                          "${supportPath.path}/${ClassTableController.decorationName}",
-                        );
-                        preference.setBool(
-                          preference.Preference.decoration,
-                          true,
-                        );
-                        if (context.mounted) {
-                          showToast(
-                            context: context,
-                            msg: FlutterI18n.translate(
-                              context,
-                              "setting.successful_setting",
-                            ),
-                          );
-                        }
-                      } else {
-                        if (context.mounted) {
-                          showToast(
-                            context: context,
-                            msg: FlutterI18n.translate(
-                              context,
-                              "setting.failure_setting",
-                            ),
-                          );
-                        }
-                      }
-                    }
-                  },
-                ),
-                const Divider(),
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(context, "setting.clear_user_class"),
-                  ),
-                  trailing: const Icon(Icons.navigate_next),
-                  onTap: () => showDialog<String>(
-                    context: context,
-                    builder: (BuildContext context) => AlertDialog(
-                      title: Text(
-                        FlutterI18n.translate(
-                          context,
-                          "setting.clear_user_class_title",
-                        ),
-                      ),
-                      content: Text(
-                        FlutterI18n.translate(
-                          context,
-                          "setting.clear_user_class_content",
-                        ),
-                      ),
-                      actions: [
-                        TextButton(
-                          style: TextButton.styleFrom(
-                            backgroundColor: Theme.of(
-                              context,
-                            ).colorScheme.primary,
-                            foregroundColor: Theme.of(
-                              context,
-                            ).colorScheme.onPrimary,
-                          ),
-                          onPressed: () => Navigator.pop(context),
-                          child: Text(FlutterI18n.translate(context, "cancel")),
-                        ),
-                        TextButton(
-                          onPressed: () {
-                            UserDefinedClassFile.clearUserDefinedClass();
-                            ClassTableController
-                                    .i
-                                    .userDefinedClassSignal
-                                    .value =
-                                UserDefinedClassData.empty();
-                            if (mounted) {
-                              setState(() {});
-                            }
-                            showToast(
-                              context: context,
-                              msg: FlutterI18n.translate(
-                                context,
-                                "setting.clear_user_class_clear",
-                              ),
-                            );
-                            Navigator.pop(context);
-                          },
-                          child: Text(
-                            FlutterI18n.translate(context, "confirm"),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-                const Divider(),
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(context, "setting.class_refresh"),
-                  ),
-                  trailing: const Icon(Icons.navigate_next),
-                  onTap: () => showDialog<String>(
-                    context: context,
-                    builder: (BuildContext context) => AlertDialog(
-                      title: Text(
-                        FlutterI18n.translate(
-                          context,
-                          "setting.class_refresh_title",
-                        ),
-                      ),
-                      content: Text(
-                        FlutterI18n.translate(
-                          context,
-                          "setting.class_refresh_content",
-                        ),
-                      ),
-                      actions: [
-                        TextButton(
-                          style: TextButton.styleFrom(
-                            backgroundColor: Theme.of(
-                              context,
-                            ).colorScheme.primary,
-                            foregroundColor: Theme.of(
-                              context,
-                            ).colorScheme.onPrimary,
-                          ),
-                          onPressed: () => Navigator.pop(context),
-                          child: Text(FlutterI18n.translate(context, "cancel")),
-                        ),
-                        TextButton(
-                          onPressed: () async {
-                            await Future.wait([
-                              ClassTableController.i.reloadClassTable(),
-                              ExamController.i.reloadExamInfo(),
-                              PhysicsExperimentController.i
-                                  .reloadPhysicsExperiment(),
-                              OtherExperimentController.i
-                                  .reloadOtherExperiment(),
-                            ]);
-                            await maybeAutoSyncSystemCalendar();
-                            if (mounted) {
-                              setState(() {});
-                            }
-                            if (context.mounted) {
-                              Navigator.pop(context);
-                            }
-                          },
-                          child: Text(
-                            FlutterI18n.translate(context, "confirm"),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-                const Divider(),
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(context, "setting.class_swift"),
-                  ),
-                  subtitle: Text(
-                    FlutterI18n.translate(
-                      context,
-                      "setting.class_swift_description",
-                      translationParams: {
-                        "swift": preference
-                            .getInt(preference.Preference.swift)
-                            .toString(),
-                      },
-                    ),
-                  ),
-                  trailing: const Icon(Icons.navigate_next),
-                  onTap: () {
-                    showDialog(
-                      barrierDismissible: false,
-                      context: context,
-                      builder: (context) => ChangeSwiftDialog(),
-                    ).then((value) {
-                      setState(() {});
-                    });
-                  },
-                ),
-                const Divider(),
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(context, "setting.semester_change"),
-                  ),
-                  subtitle: Text(
-                    FlutterI18n.translate(
-                      context,
-                      "setting.semester_change_description",
-                      translationParams: {
-                        "semester": preference.getString(
-                          preference.Preference.currentSemester,
-                        ),
-                      },
-                    ),
-                  ),
-                  trailing: const Icon(Icons.navigate_next),
-                  onTap: () {
-                    showDialog<bool>(
-                      barrierDismissible: false,
-                      context: context,
-                      builder: (context) => SemesterSwitchDialog(),
-                    ).then((value) async {
-                      if (value == true) {
-                        setState(() {});
-                        if (context.mounted) {
-                          showToast(context: context, msg: "Updating data");
-                        }
-                        await _waitForSemesterAwareReloads();
-                        await maybeAutoSyncSystemCalendar();
-                        if (mounted) {
-                          setState(() {});
-                        }
-                      }
-                    });
-                  },
-                ),
-              ],
-            ),
-          ),
-          ReXCard(
-            title: _buildListSubtitle(
-              FlutterI18n.translate(context, "setting.core_setting"),
-            ),
-            remaining: const [],
-            bottomRow: Column(
-              children: [
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(context, "setting.check_logger"),
-                  ),
-                  trailing: const Icon(Icons.navigate_next),
-                  onTap: () => context.push(TalkerScreen(talker: log)),
-                ),
-                const Divider(),
-                if (Platform.isAndroid || Platform.isIOS) ...[
-                  ListTile(
-                    title: Text(
-                      FlutterI18n.translate(
-                        context,
-                        "setting.notification_debug_page",
-                      ),
-                    ),
-                    trailing: const Icon(Icons.navigate_next),
-                    onTap: () => context.push(NotificationDebugPage()),
-                  ),
-                  const Divider(),
-                ],
-                ListTile(
-                  title: Text(
-                    FlutterI18n.translate(context, "setting.clear_and_restart"),
-                  ),
-                  trailing: const Icon(Icons.navigate_next),
-                  onTap: () => showDialog<String>(
-                    context: context,
-                    barrierDismissible: false,
-                    builder: (BuildContext context) => AlertDialog(
-                      title: Text(
-                        FlutterI18n.translate(
-                          context,
-                          "setting.clear_and_restart_dialog.title",
-                        ),
-                      ),
-                      content: Text(
-                        FlutterI18n.translate(
-                          context,
-                          "setting.clear_and_restart_dialog.content",
-                        ),
-                      ),
-                      actions: [
-                        TextButton(
-                          style: TextButton.styleFrom(
-                            backgroundColor: Theme.of(
-                              context,
-                            ).colorScheme.primary,
-                            foregroundColor: Theme.of(
-                              context,
-                            ).colorScheme.onPrimary,
-                          ),
-                          onPressed: () => Navigator.pop(context),
-                          child: Text(FlutterI18n.translate(context, "cancel")),
-                        ),
-                        TextButton(
-                          onPressed: () async {
-                            ProgressDialog pd = ProgressDialog(
-                              context: context,
-                            );
-                            pd.show(
-                              msg: FlutterI18n.translate(
-                                context,
-                                "setting.clear_and_restart_dialog.cleaning",
-                              ),
-                            );
-
-                            /// Clean Cookie
-                            try {
-                              await NetworkSession().clearCookieJar();
-                              // I don't care.
-                              // ignore: empty_catches
-                            } on Exception {}
-
-                            /// Clean sport cookie.
-                            try {
-                              await SportSession().sportCookieJar.deleteAll();
-                              // I don't care.
-                              // ignore: empty_catches
-                            } on Exception {}
-
-                            /// Clean cache.
-                            EnergySession.clearCache();
-                            EnergySession.clearElectricityHistory();
-                            for (var value in [
-                              ClassTableSession.schoolClassName,
-                              ExamSession.examDataCacheName,
-                              ExperimentSession.physicsExperimentCacheName,
-                              SysjSession.otherExperimentCacheName,
-                              ScoreSession.scoreListCacheName,
-                            ]) {
-                              var file = File("${supportPath.path}/$value");
-                              if (file.existsSync()) {
-                                file.deleteSync();
-                              }
-                            }
-
-                            if (context.mounted) {
-                              showToast(
-                                context: context,
-                                msg: FlutterI18n.translate(
-                                  context,
-                                  "setting.clear_and_restart_dialog.clear",
-                                ),
-                              );
-                              if (Platform.isIOS) {
-                                Restart.restartApp(
-                                  mode: RestartMode.notificationFallback,
-                                  notificationTitle: FlutterI18n.translate(
-                                    context,
-                                    "restart_app.title_cache_cleared",
-                                  ),
-                                  notificationBody: FlutterI18n.translate(
-                                    context,
-                                    "restart_app.content",
-                                  ),
-                                );
-                              } else {
-                                Restart.restartApp();
-                              }
-                            }
-                          },
-                          child: Text(
-                            FlutterI18n.translate(context, "confirm"),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-                const Divider(),
-                ListTile(
-                  title: Text(FlutterI18n.translate(context, "setting.logout")),
-                  trailing: const Icon(Icons.navigate_next),
-                  onTap: () => showDialog<String>(
-                    context: context,
-                    barrierDismissible: false,
-                    builder: (BuildContext context) => AlertDialog(
-                      title: Text(
-                        FlutterI18n.translate(
-                          context,
-                          "setting.logout_dialog.title",
-                        ),
-                      ),
-                      content: Text(
-                        FlutterI18n.translate(
-                          context,
-                          "setting.logout_dialog.content",
-                        ),
-                      ),
-                      actions: [
-                        TextButton(
-                          style: TextButton.styleFrom(
-                            backgroundColor: Theme.of(
-                              context,
-                            ).colorScheme.primary,
-                            foregroundColor: Theme.of(
-                              context,
-                            ).colorScheme.onPrimary,
-                          ),
-                          onPressed: () => Navigator.pop(context),
-                          child: Text(FlutterI18n.translate(context, "cancel")),
-                        ),
-                        TextButton(
-                          onPressed: () async {
-                            ProgressDialog pd = ProgressDialog(
-                              context: context,
-                            );
-                            pd.show(
-                              msg: FlutterI18n.translate(
-                                context,
-                                "setting.logout_dialog.logging_out",
-                              ),
-                            );
-
-                            /// Clean Cookie
-                            try {
-                              await NetworkSession().clearCookieJar();
-                              // I don't care.
-                              // ignore: empty_catches
-                            } on Exception {}
-
-                            /// Clean sport cookie.
-                            try {
-                              await SportSession().sportCookieJar.deleteAll();
-                              // I don't care.
-                              // ignore: empty_catches
-                            } on Exception {}
-
-                            /// Clean all.
-                            EnergySession.clearCache();
-                            EnergySession.clearElectricityHistory();
-                            for (var value in [
-                              ClassTableSession.schoolClassName,
-                              UserDefinedClassFile.userDefinedClassName,
-                              ClassTableController.decorationName,
-                              ExamSession.examDataCacheName,
-                              ExperimentSession.physicsExperimentCacheName,
-                              SysjSession.otherExperimentCacheName,
-                              ScoreSession.scoreListCacheName,
-                            ]) {
-                              var file = File("${supportPath.path}/$value");
-                              if (file.existsSync()) {
-                                file.deleteSync();
-                              }
-                            }
-                            await GetIt.instance<RuisiService>().logout();
-
-                            /// Clean user information
-                            await preference.prefrenceClear();
-
-                            /// Theme back to default
-                            ThemeController.i.updateTheme();
-
-                            /// Sync widget login state
-                            await syncWidgetLoginState(false);
-
-                            /// Clean iOS widget data files
-                            await clearWidgetFiles();
-
-                            /// Restart app
-                            if (context.mounted) {
-                              pd.close();
-                              if (Platform.isIOS) {
-                                Restart.restartApp(
-                                  mode: RestartMode.notificationFallback,
-                                  notificationTitle: FlutterI18n.translate(
-                                    context,
-                                    "restart_app.title_logged_out",
-                                  ),
-                                  notificationBody: FlutterI18n.translate(
-                                    context,
-                                    "restart_app.content",
-                                  ),
-                                );
-                              } else {
-                                Restart.restartApp();
-                              }
-                            }
-                          },
-                          child: Text(
-                            FlutterI18n.translate(context, "confirm"),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
+          const SettingAboutSection(),
+          SettingUiSection(onChanged: _refresh),
+          const SettingAccountSection(),
+          const SettingNotificationSection(),
+          SettingClassTableSection(actions: _actions, onChanged: _refresh),
+          SettingCoreSection(actions: _actions),
         ],
       ).constrained(maxWidth: 600).center().safeArea(top: true),
     );
