+++ lib/repository/xidian_ids/learning_session.dart
@@ -11,7 +11,7 @@ import 'package:html/dom.dart';
 import 'package:html/parser.dart';
 import 'package:watermeter/model/xidian_ids/class_attendance.dart';
 import 'package:watermeter/repository/logger.dart';
-import 'package:watermeter/page/login/jc_captcha.dart';
+import 'package:watermeter/repository/xidian_ids/slider_captcha_client.dart';
 import 'package:watermeter/repository/xidian_ids/ids_session.dart';
 
 class LearningSession extends IDSSession {
@@ -45,7 +45,7 @@ class LearningSession extends IDSSession {
     String? location = await checkAndLogin(
       target: LOGIN_URL,
       sliderCaptcha: (String cookieStr) =>
-          SliderCaptchaClientProvider(cookie: cookieStr).solve(null),
+          SliderCaptchaClientProvider(cookie: cookieStr).solve(),
     );
 
     while (location != null) {
